refactor(General): :recycle: Disable metric flag
