+++ config/webpack.common.js
@@ -39,7 +39,8 @@ module.exports = {
       'process.env.BRAND_APP_NAME': JSON.stringify(process.env.BRAND_APP_NAME),
       'process.env.BRAND_APP_LOGO': JSON.stringify(process.env.BRAND_APP_LOGO || ''),
       'process.env.APP_VERSION': JSON.stringify(process.env.APP_VERSION) || JSON.stringify(version),
-      'process.env.COLLECTOR_URL': JSON.stringify(process.env.COLLECTOR_URL || '')
+      'process.env.COLLECTOR_URL': JSON.stringify(process.env.COLLECTOR_URL || ''),
+      'process.env.DISABLE_METRICS': JSON.stringify(process.env.DISABLE_METRICS)
     }),
     new HtmlWebpackPlugin({
       template: path.join(ROOT, '/public/index.html'),
